// src/types.ts
export enum DataSource {
    FILE = 'file',
    DATABASE = 'database',
    API = 'api',
    STREAM = 'stream',
    MONGODB = 'mongodb',
    XML = 'xml',
    PIPELINE = 'pipeline'
}

export enum DataFormat {
    JSON = 'json',
    CSV = 'csv',
    XML = 'xml',
    YAML = 'yaml',
    TEXT = 'text',
    MONGODB = 'mongodb',
    PIPELINE = 'pipeline'
}
// ...
export class DataError extends Error {
    constructor(
        message: string,
        public source: DataSource | DataFormat,
        public code: string,
        public details?: unknown
    ) {
        super(message);
        this.name = 'DataError';
    }
}